backend/requirements.txt
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
assemblyai==0.30.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
dnspython==2.7.0
fastapi==0.115.4
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-generativeai==0.8.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx>=0.25.0,<=0.27.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core>=0.3.17,<0.4.0
langsmith==0.1.140
loguru==0.7.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy>=1.26.0,<2.0.0
packaging==24.2
requests-toolbelt==1.0.0
weaviate-client==4.9.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy>=1.4,<2.0.36
starlette==0.41.0
tenacity==9.0.0
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
